Upgrade to Druid 0.15 and Spark 2.4 with dataframe api
