+++ project/build.properties
@@ -17,4 +17,4 @@
 #  under the License.
 #
 
-sbt.version = 0.13.8
+sbt.version = 0.13.18
