+++ src/main/resources/META-INF/services/org.apache.druid.initialization.DruidModule
@@ -16,4 +16,4 @@
 # specific language governing permissions and limitations
 # under the License.
 #
-io.druid.indexer.spark.SparkDruidIndexerModule
+org.apache.druid.indexer.spark.SparkDruidIndexerModule
