+++ src/test/resources/text/event.json
