+++ src/test/resources/text/lineitem.small.tbl
