+++ src/test/scala/org/apache/druid/indexer/spark/TestSparkDruidIndexer.scala
@@ -17,69 +17,43 @@
  *  under the License.
  */
 
-package io.druid.indexer.spark
+package org.apache.druid.indexer.spark
 
 import java.io.{Closeable, File}
 import java.nio.file.Files
 import java.util
-import com.google.common.collect.ImmutableList
+
 import com.google.common.io.Closer
-import io.druid.data.input.impl.{DimensionsSpec, JSONParseSpec, StringDimensionSchema, TimestampSpec}
-import io.druid.java.util.common.JodaUtils
-import io.druid.java.util.common.granularity.Granularities
-import io.druid.java.util.common.logger.Logger
-import io.druid.java.util.common.{CompressionUtils, IAE}
-import io.druid.query.aggregation.LongSumAggregatorFactory
-import io.druid.segment.QueryableIndexIndexableAdapter
 import org.apache.commons.io.FileUtils
-import org.apache.spark.{SparkConf, SparkContext}
+import org.apache.druid.data.input.impl._
+import org.apache.druid.data.input.parquet.avro.ParquetAvroHadoopInputRowParser
+import org.apache.druid.data.input.parquet.simple.ParquetParseSpec
+import org.apache.druid.java.util.common.granularity.Granularities
+import org.apache.druid.java.util.common.logger.Logger
+import org.apache.druid.java.util.common.{IAE, JodaUtils}
+import org.apache.druid.query.aggregation.LongSumAggregatorFactory
+import org.apache.druid.segment.QueryableIndexIndexableAdapter
+import org.apache.druid.segment.column.NumericColumn
+import org.apache.druid.utils.CompressionUtils
+import org.apache.spark.SparkConf
+import org.apache.spark.sql.SparkSession
 import org.joda.time.{DateTime, Interval}
 import org.scalatest._
+
 import scala.collection.JavaConverters._
 
+class TestSparkDruidIndexer extends FlatSpec with Matchers {
 
-class TestSparkDruidIndexer extends FlatSpec with Matchers
-{
-
-  import TestScalaBatchIndexTask._
+  import TestConfProvider._
 
   "The spark indexer" should "return proper DataSegments" in {
     val data_files = Seq(
-      this.getClass.getResource("/lineitem.small.tbl").toString,
-      this.getClass.getResource("/empty.tbl").toString
+      this.getClass.getResource("/text/lineitem.small.tbl").getPath,
+      this.getClass.getResource("/text/empty.tbl").getPath
     )
-    val closer = Closer.create()
     val outDir = Files.createTempDirectory("segments").toFile
-    (outDir.mkdirs() || outDir.exists()) && outDir.isDirectory should be(true)
-    closer.register(
-      new Closeable()
-      {
-        override def close(): Unit = FileUtils.deleteDirectory(outDir)
-      }
-    )
+    val closer = Closer.create()
     try {
-      val conf = new SparkConf()
-        .setAppName("Simple Application")
-        .setMaster("local[4]")
-        .set("user.timezone", "UTC")
-        .set("file.encoding", "UTF-8")
-        .set("java.util.logging.manager", "org.apache.logging.log4j.jul.LogManager")
-        .set("org.jboss.logging.provider", "slf4j")
-        .set("druid.processing.columnCache.sizeBytes", "1000000000")
-        .set("spark.driver.host", "localhost")
-        .set("spark.executor.userClassPathFirst", "true")
-        .set("spark.driver.userClassPathFirst", "true")
-        .set("spark.kryo.referenceTracking", "false")
-        .registerKryoClasses(SparkBatchIndexTask.getKryoClasses())
-
-      val sc = new SparkContext(conf)
-      closer.register(
-        new Closeable
-        {
-          override def close(): Unit = sc.stop()
-        }
-      )
-
       val loadResults = SparkDruidIndexer.loadData(
         data_files,
         new SerializedJson(dataSchema),
@@ -89,7 +63,7 @@ class TestSparkDruidIndexer extends FlatSpec with Matchers
         outDir.toString,
         indexSpec,
         buildV9Directly,
-        sc
+        buildSparkSession(outDir, closer)
       )
       loadResults.length should be(7)
       for (
@@ -108,7 +82,7 @@ class TestSparkDruidIndexer extends FlatSpec with Matchers
         segment.getMetrics.asScala.toList should equal(dataSchema.getAggregators.map(_.getName).toList)
         val file = new File(segment.getLoadSpec.get("path").toString)
         file.exists() should be(true)
-        val segDir = Files.createTempDirectory(outDir.toPath, "loadableSegment-%s" format segment.getIdentifier).toFile
+        val segDir = Files.createTempDirectory(outDir.toPath, "loadableSegment-%s" format segment.getId).toFile
         val copyResult = CompressionUtils.unzip(file, segDir)
         copyResult.size should be > 0L
         copyResult.getFiles.asScala.map(_.getName).toSet should equal(
@@ -116,18 +90,18 @@ class TestSparkDruidIndexer extends FlatSpec with Matchers
         )
         val index = StaticIndex.INDEX_IO.loadIndex(segDir)
         try {
-          val qindex = new QueryableIndexIndexableAdapter(index)
-          qindex.getDimensionNames.asScala.toSet should
+          val indexAdapter = new QueryableIndexIndexableAdapter(index)
+          indexAdapter.getDimensionNames.asScala.toSet should
             equal(
               dataSchema.getParser.getParseSpec.getDimensionsSpec.getDimensionNames.asScala.toSet --
                 dataSchema.getParser.getParseSpec.getDimensionsSpec.getDimensionExclusions.asScala.toSet
             )
-          for (dimension <- qindex.getDimensionNames.iterator().asScala) {
-            val dimVal = qindex.getDimValueLookup(dimension).asScala
+          for (dimension <- indexAdapter.getDimensionNames.iterator().asScala) {
+            val dimVal = indexAdapter.getDimValueLookup[String](dimension).asScala
             dimVal should not be 'Empty
             for (dv <- dimVal) {
               Option(dv) match {
-                case Some(v) =>
+                case Some(_) =>
                   // I had a problem at one point where dimension values were being stored as lists
                   // This is a check to make sure the dimension is a list of values rather than being a list of lists
                   // If the unit test is ever modified to have dimension values that start with this offending case
@@ -138,11 +112,11 @@ class TestSparkDruidIndexer extends FlatSpec with Matchers
               }
             }
           }
-          qindex.getNumRows should be > 0
+          indexAdapter.getNumRows should be > 0
           for (colName <- Seq("count")) {
-            val column = index.getColumn(colName).getGenericColumn
+            val column = index.getColumnHolder(colName).getColumn.asInstanceOf[NumericColumn]
             try {
-              for (i <- Range.apply(0, qindex.getNumRows)) {
+              for (i <- Range.apply(0, indexAdapter.getNumRows)) {
                 column.getLongSingleValueRow(i) should not be 0
               }
             }
@@ -151,23 +125,24 @@ class TestSparkDruidIndexer extends FlatSpec with Matchers
             }
           }
           for (colName <- Seq("L_QUANTITY_longSum")) {
-            val column = index.getColumn(colName).getGenericColumn
+            val column = index.getColumnHolder(colName).getColumn.asInstanceOf[NumericColumn]
             try {
-              Range.apply(0, qindex.getNumRows).map(column.getLongSingleValueRow).sum should not be 0
+              Range.apply(0, indexAdapter.getNumRows).map(column.getLongSingleValueRow).sum should not be 0
             }
             finally {
               column.close()
             }
           }
-          for (colName <- Seq("L_DISCOUNT_doubleSum", "L_TAX_doubleSum")) {
-            val column = index.getColumn(colName).getGenericColumn
+          /*for (colName <- Seq("L_DISCOUNT_doubleSum", "L_TAX_doubleSum")) {
+            val columnHolder = index.getColumnHolder(colName)
+            val valueSelector = columnHolder.makeNewSettableColumnValueSelector()
             try {
-              Range.apply(0, qindex.getNumRows).map(column.getFloatSingleValueRow).sum should not be 0.0D
+              Range.apply(0, indexAdapter.getNumRows).map(_ => valueSelector.getDouble).sum should not be 0.0D
             }
             finally {
-              column.close()
+              columnHolder.getColumn.close()
             }
-          }
+          }*/
           index.getDataInterval.getEnd.getMillis should not be JodaUtils.MAX_INSTANT
         }
         finally {
@@ -180,53 +155,26 @@ class TestSparkDruidIndexer extends FlatSpec with Matchers
     }
   }
 
-
   it should "return proper DataSegments from json" in {
-    val data_files = Seq(this.getClass.getResource("/event.json").toString)
+    val data_files = Seq(this.getClass.getResource("/text/event.json").toString)
     val closer = Closer.create()
     val outDir = Files.createTempDirectory("segments").toFile
-    (outDir.mkdirs() || outDir.exists()) && outDir.isDirectory should be(true)
-    closer.register(
-      new Closeable()
-      {
-        override def close(): Unit = FileUtils.deleteDirectory(outDir)
-      }
-    )
     try {
-      val conf = new SparkConf()
-        .setAppName("Simple Application")
-        .setMaster("local[4]")
-        .set("user.timezone", "UTC")
-        .set("file.encoding", "UTF-8")
-        .set("java.util.logging.manager", "org.apache.logging.log4j.jul.LogManager")
-        .set("org.jboss.logging.provider", "slf4j")
-        .set("druid.processing.columnCache.sizeBytes", "1000000000")
-        .set("spark.driver.host", "localhost")
-        .set("spark.executor.userClassPathFirst", "true")
-        .set("spark.driver.userClassPathFirst", "true")
-        .set("spark.kryo.referenceTracking", "false")
-        .registerKryoClasses(SparkBatchIndexTask.getKryoClasses())
-
-      val sc = new SparkContext(conf)
-      closer.register(
-        new Closeable
-        {
-          override def close(): Unit = sc.stop()
-        }
-      )
       val aggName = "agg_met1"
       val aggregatorFactory = new LongSumAggregatorFactory(aggName, "met1")
       val dataSchema = buildDataSchema(
-        parseSpec = new
-            JSONParseSpec(
-              new TimestampSpec("ts", null, null),
-              new DimensionsSpec(ImmutableList.of(new StringDimensionSchema("dim1")), ImmutableList.of("ts"), null),
-              null,
-              null
-            ),
+        parseSpec = new JSONParseSpec(
+          new TimestampSpec("ts", null, null),
+          new DimensionsSpec(
+            Seq(new StringDimensionSchema("dim1").asInstanceOf[DimensionSchema]).asJava,
+            Seq("ts").asJava,
+            null
+          ),
+          null,
+          null
+        ),
         aggFactories = Seq(aggregatorFactory)
       )
-
       val loadResults = SparkDruidIndexer.loadData(
         data_files,
         new SerializedJson(dataSchema),
@@ -236,7 +184,7 @@ class TestSparkDruidIndexer extends FlatSpec with Matchers
         outDir.toString,
         indexSpec,
         buildV9Directly,
-        sc
+        buildSparkSession(outDir, closer)
       )
       loadResults.length should be(1)
       for (
@@ -251,7 +199,7 @@ class TestSparkDruidIndexer extends FlatSpec with Matchers
         segment.getMetrics.asScala.toList should equal(dataSchema.getAggregators.map(_.getName).toList)
         val file = new File(segment.getLoadSpec.get("path").toString)
         file.exists() should be(true)
-        val segDir = Files.createTempDirectory(outDir.toPath, "loadableSegment-%s" format segment.getIdentifier).toFile
+        val segDir = Files.createTempDirectory(outDir.toPath, "loadableSegment-%s" format segment.getId).toFile
         val copyResult = CompressionUtils.unzip(file, segDir)
         copyResult.size should be > 0L
         copyResult.getFiles.asScala.map(_.getName).toSet should equal(
@@ -261,7 +209,7 @@ class TestSparkDruidIndexer extends FlatSpec with Matchers
         try {
           val qindex = new QueryableIndexIndexableAdapter(index)
           qindex.getDimensionNames.asScala.toSet should equal(Set("dim1"))
-          val dimVal = qindex.getDimValueLookup("dim1").asScala
+          val dimVal : Iterable[String] = qindex.getDimValueLookup("dim1").asScala
           dimVal should not be 'Empty
           for (dv <- dimVal) {
             dv should equal("val1")
@@ -269,7 +217,82 @@ class TestSparkDruidIndexer extends FlatSpec with Matchers
           qindex.getMetricNames.asScala.toSet should equal(Set(aggName))
           qindex.getMetricType(aggName) should equal(aggregatorFactory.getTypeName)
           qindex.getNumRows should be(1)
-          qindex.getRows.asScala.head.getMetrics()(0) should be(1)
+          index.getDataInterval.getEnd.getMillis should not be JodaUtils.MAX_INSTANT
+        }
+        finally {
+          index.close()
+        }
+      }
+    }
+    finally {
+      closer.close()
+    }
+  }
+
+  it should "return proper DataSegments from parquet" in {
+    val dataFiles = Seq(this.getClass.getResource("/parquet/event.parquet").toString)
+    val closer = Closer.create()
+    val outDir = Files.createTempDirectory("segments").toFile
+    try {
+      val aggName = "agg_met1"
+      val aggregatorFactory = new LongSumAggregatorFactory(aggName, "met1")
+      val dataSchema = buildDataSchema(
+        parseSpec = new
+          ParquetParseSpec(
+            new TimestampSpec("ts", null, null),
+            new DimensionsSpec(
+              Seq(new StringDimensionSchema("dim1").asInstanceOf[DimensionSchema]).asJava,
+              Seq("ts").asJava,
+              null
+            ),
+            null
+          ),
+        parser = spec => new ParquetAvroHadoopInputRowParser(spec, false),
+        aggFactories = Seq(aggregatorFactory)
+      )
+
+      val loadResults = SparkDruidIndexer.loadData(
+        dataFiles,
+        new SerializedJson(dataSchema),
+        SparkBatchIndexTask.mapToSegmentIntervals(Seq(interval), Granularities.YEAR),
+        rowsPerPartition,
+        rowsPerFlush,
+        outDir.toString,
+        indexSpec,
+        buildV9Directly,
+        buildSparkSession(outDir, closer)
+      )
+      loadResults.length should be(1)
+      for (
+        segment <- loadResults
+      ) {
+        segment.getBinaryVersion should be(9)
+        segment.getDataSource should equal(dataSource)
+        interval.contains(segment.getInterval) should be(true)
+        segment.getInterval.contains(interval) should be(false)
+        segment.getSize should be > 0L
+        segment.getDimensions.asScala.toSet should equal(Set("dim1"))
+        segment.getMetrics.asScala.toList should equal(dataSchema.getAggregators.map(_.getName).toList)
+        val file = new File(segment.getLoadSpec.get("path").toString)
+        file.exists() should be(true)
+        val segDir = Files.createTempDirectory(outDir.toPath, "loadableSegment-%s" format segment.getId.toString).toFile
+        val copyResult = CompressionUtils.unzip(file, segDir)
+        copyResult.size should be > 0L
+        copyResult.getFiles.asScala.map(_.getName).toSet should equal(
+          Set("00000.smoosh", "meta.smoosh", "version.bin", "factory.json")
+        )
+        val index = StaticIndex.INDEX_IO.loadIndex(segDir)
+        try {
+          val qIndex = new QueryableIndexIndexableAdapter(index)
+          qIndex.getDimensionNames.asScala.toSet should equal(Set("dim1"))
+          val dimVal = qIndex.getDimValueLookup[String]("dim1").asScala
+          dimVal should not be 'Empty
+          for (dv <- dimVal) {
+            dv should equal("val1")
+          }
+          qIndex.getMetricNames.asScala.toSet should equal(Set(aggName))
+          qIndex.getMetricType(aggName) should equal(aggregatorFactory.getTypeName)
+          qIndex.getNumRows should be(1)
           index.getDataInterval.getEnd.getMillis should not be JodaUtils.MAX_INSTANT
         }
         finally {
@@ -352,18 +375,17 @@ class TestSparkDruidIndexer extends FlatSpec with Matchers
     partitioner.getPartition(makeEvent(intervals.last.getEnd.minus(10L), "anotherHash1")) should equal(1)
   }
 
-  it should "properly partition multiple date ranges and buckets when dim is specified" in
-    {
-      val intervals = SparkBatchIndexTask.mapToSegmentIntervals(Seq(Interval.parse("1992/1994")), Granularities.YEAR)
-      val map = Map(new DateTime("1992").getMillis -> 100L, new DateTime("1993").getMillis -> 200L)
-      val m = SparkDruidIndexer.getSizedPartitionMap(map, 150)
-      val partitioner = new DateBucketAndHashPartitioner(Granularities.YEAR, m, Option(Set("dim1")))
-      partitioner.getPartition(makeEvent(intervals.head.getStart)) should equal(0)
-      partitioner.getPartition(makeEvent(intervals.last.getEnd.minus(10L))) should equal(1)
-      partitioner.getPartition(makeEvent(intervals.last.getEnd.minus(10L), "something else1")) should equal(2)
-      partitioner.getPartition(makeEvent(intervals.last.getEnd.minus(10L), "anotherHash3")) should equal(2)
-      partitioner.getPartition(makeEvent(intervals.last.getEnd.minus(10L), "anotherHash1")) should equal(1)
-    }
+  it should "properly partition multiple date ranges and buckets when dim is specified" in {
+    val intervals = SparkBatchIndexTask.mapToSegmentIntervals(Seq(Interval.parse("1992/1994")), Granularities.YEAR)
+    val map = Map(new DateTime("1992").getMillis -> 100L, new DateTime("1993").getMillis -> 200L)
+    val m = SparkDruidIndexer.getSizedPartitionMap(map, 150)
+    val partitioner = new DateBucketAndHashPartitioner(Granularities.YEAR, m, Option(Set("dim1")))
+    partitioner.getPartition(makeEvent(intervals.head.getStart)) should equal(0)
+    partitioner.getPartition(makeEvent(intervals.last.getEnd.minus(10L))) should equal(1)
+    partitioner.getPartition(makeEvent(intervals.last.getEnd.minus(10L), "something else1")) should equal(2)
+    partitioner.getPartition(makeEvent(intervals.last.getEnd.minus(10L), "anotherHash3")) should equal(2)
+    partitioner.getPartition(makeEvent(intervals.last.getEnd.minus(10L), "anotherHash1")) should equal(1)
+  }
 
   it should "properly group multiple events together" in {
     val intervals = SparkBatchIndexTask.mapToSegmentIntervals(Seq(Interval.parse("1992/1994")), Granularities.YEAR)
@@ -437,9 +459,40 @@ class TestSparkDruidIndexer extends FlatSpec with Matchers
       }
     ) should be(2)
   }
+
+  private def buildSparkSession(outDir: File, closer: Closer): SparkSession = {
+    (outDir.mkdirs() || outDir.exists()) && outDir.isDirectory should be(true)
+    closer.register(
+      new Closeable() {
+        override def close(): Unit = FileUtils.deleteDirectory(outDir)
+      }
+    )
+
+    val conf = new SparkConf()
+      .setAppName("Simple Application")
+      .setMaster("local[*]")
+      .set("user.timezone", "UTC")
+      .set("file.encoding", "UTF-8")
+      .set("java.util.logging.manager", "org.apache.logging.log4j.jul.LogManager")
+      .set("org.jboss.logging.provider", "slf4j")
+      .set("druid.processing.columnCache.sizeBytes", "1000000000")
+      .set("spark.driver.host", "localhost")
+      .set("spark.executor.userClassPathFirst", "true")
+      .set("spark.driver.userClassPathFirst", "true")
+      .set("spark.kryo.referenceTracking", "false")
+      .set("spark.kryo.registrator", "org.apache.druid.indexer.spark.SparkDruidRegistrator")
+      .registerKryoClasses(SparkBatchIndexTask.getKryoClasses())
+
+    val sparkSession = SparkSession.builder.config(conf).getOrCreate()
+    closer.register(
+      new Closeable {
+        override def close(): Unit = sparkSession.stop()
+      }
+    )
+    sparkSession
+  }
 }
 
-object StaticTestSparkDruidIndexer
-{
+object StaticTestSparkDruidIndexer {
   val log = new Logger(classOf[TestSparkDruidIndexer])
 }
