+++ version.sbt
@@ -1,1 +1,1 @@
-version in ThisBuild := "0.12.1.1-SNAPSHOT"
+version in ThisBuild := "0.15.1-incubating"
